requirements.txt
MetaTrader5==5.0.5430
pandas
numpy
python-dotenv
loguru
openai
google-generativeai
feedparser
requests
